requirements.txt
flask==2.0.1
requests==2.27.1
python-dotenv==0.20.0
moviepy==1.0.3
numpy==1.22.4
Pillow==9.1.1
werkzeug==2.0.3
gunicorn==20.1.0
celery==5.2.7
redis==4.3.4
boto3==1.24.28 